int64 for all IDs

+++ browser.go
@@ -19,7 +19,7 @@ const (
 	TWITTER                   // 12
 )
 
-var Browsers = map[uint]string{
+var Browsers = map[int64]string{
 	UNKNOWN:        "Unknown",
 	IE:             "IE",
 	CHROME:         "Chrome",
@@ -198,6 +198,6 @@ var browsers = []pattern{
 
 // Browser will return the browser id and version number.
 // Returns 1,0 when nothing matched.
-func Browser(userAgent string) (uint, int, int) {
+func Browser(userAgent string) (int64, int, int) {
 	return find(browsers, userAgent)
 }
